mvoe pauli_eigs
