+++ pennylane/pauli/__init__.py
@@ -34,6 +34,7 @@ from .utils import (
     diagonalize_qwc_pauli_words,
     diagonalize_qwc_groupings,
     simplify,
+    pauli_eigs,
 )
 
 from .pauli_interface import pauli_word_prefactor
