+++ tests/pauli/test_pauli_utils.py
@@ -20,6 +20,9 @@ import warnings
 import numpy as np
 import pytest
 
+import functools
+import itertools
+
 import pennylane as qml
 from pennylane import (
     RX,
@@ -45,6 +48,7 @@ from pennylane.pauli import (
     is_qwc,
     observables_to_binary_matrix,
     partition_pauli_group,
+    pauli_eigs,
     pauli_group,
     pauli_to_binary,
     pauli_word_to_matrix,
@@ -1135,3 +1139,41 @@ class TestTapering:
         pws_lst = [list(qml.pauli.pauli_sentence(t))[0] for t in terms]
         binary_matrix = qml.pauli.utils._binary_matrix_from_pws(pws_lst, num_qubits)
         assert (binary_matrix == result).all()
+
+
+class TestPauliEigs:
+    """Tests for the auxiliary function to return the eigenvalues for Paulis"""
+
+    paulix = np.array([[0, 1], [1, 0]])
+    pauliy = np.array([[0, -1j], [1j, 0]])
+    pauliz = np.array([[1, 0], [0, -1]])
+    hadamard = 1 / np.sqrt(2) * np.array([[1, 1], [1, -1]])
+
+    standard_observables = [paulix, pauliy, pauliz, hadamard]
+
+    matrix_pairs = [
+        np.kron(x, y)
+        for x, y in list(itertools.product(standard_observables, standard_observables))
+    ]
+
+    def test_correct_eigenvalues_paulis(self):
+        """Test the paulieigs function for one qubit"""
+        assert np.array_equal(pauli_eigs(1), np.diag(self.pauliz))
+
+    def test_correct_eigenvalues_pauli_kronecker_products_two_qubits(self):
+        """Test the paulieigs function for two qubits"""
+        assert np.array_equal(pauli_eigs(2), np.diag(np.kron(self.pauliz, self.pauliz)))
+
+    def test_correct_eigenvalues_pauli_kronecker_products_three_qubits(self):
+        """Test the paulieigs function for three qubits"""
+        assert np.array_equal(
+            pauli_eigs(3),
+            np.diag(np.kron(self.pauliz, np.kron(self.pauliz, self.pauliz))),
+        )
+
+    @pytest.mark.parametrize("depth", list(range(1, 6)))
+    def test_cache_usage(self, depth):
+        """Test that the right number of cachings have been executed after clearing the cache"""
+        pauli_eigs.cache_clear()
+        pauli_eigs(depth)
+        assert functools._CacheInfo(depth - 1, depth, 128, depth) == pauli_eigs.cache_info()
